Reworked API + C# Update.xml
:art: API now has support for output types.

- Plain text
- JSON
- XML

These are accesible using the '?type=' parameter in the
'generate_insult.php'. For convience and backward compability the
default is plain text.

Add update.xml for the upcoming .Net Desktop Client. The file
controls/indicates the client that there is a newer version and where to
grab the corresponding file.